Big Changes: Generalization, WAD Stuff, Fixes ++++ * Support for more generalization of line types. * Add more MAPINFO details. * Add nice names to MapObjects * Fix sound origins of sectors and sector bounding boxes. * Rebuild ReMooD.WAD

+++ src/d_player.h
@@ -138,7 +138,7 @@ typedef struct player_s
 	
 	// Power ups. invinc and invis are tic counters.
 	int powers[NUMPOWERS];
-	uint8_t cards;				// bit field see declration of card_t
+	uint32_t cards;				// bit field see declration of card_t
 	bool_t backpack;
 	
 	// Frags, kills of other players.
